backend/requirements.txt
# Core Flask
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-SocketIO==5.3.5

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0

# Payment Processing
stripe==7.8.0
requests==2.31.0  # For M-Pesa Daraja API

# Real-time
python-socketio==5.10.0
eventlet==0.33.3

# Image Upload
cloudinary==1.38.0

# Utilities
python-dateutil==2.8.2
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# Development
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0

# Production Server
gunicorn==21.2.0
